fix doc build
