+++ tools/documentation_developer/CI_documentation.rst
@@ -44,7 +44,7 @@ In consequence, for testing a modification of the actual workflow or a new workf
 it requires creating a pull request, which is modified .github/workflow/ci_test.yml for
 allowing the running of this modification. You need to add the label `testCI`
 to automatically trigger the test of the CI. You can also trigger the workflow
-using with the [HTTP POST request](https://docs.github.com/en/actions/reference/workflows-and-actions/events-that-trigger-workflows#repository_dispatch).
+using with the `HTTP POST request <https://docs.github.com/en/actions/reference/workflows-and-actions/events-that-trigger-workflows#repository_dispatch>`_.
 
 Once this modification is merged into main, it should be important to clean ci_test.yml
 for having an empty workflow.
